radix i lv u <3

+++ libraries/42_libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: vkist-si <vkist-si@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/21 14:46:04 by coder             #+#    #+#             */
-/*   Updated: 2022/11/08 22:13:10 by vkist-si         ###   ########.fr       */
+/*   Updated: 2022/11/12 00:14:34 by vkist-si         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ typedef struct s_list
 typedef struct s_element
 {
 	int					content;
-	int					index;
+	long long			index;
 	int					stack_pos;
 	struct s_element	*next;
 }	t_element;
